added: v1

+++ weatherapp/Models/Minutely.swift
@@ -4,7 +4,7 @@ import Foundation
 struct Minutely: Codable {
 
   var dt            : Int? = nil
-  var precipitation : Int? = nil
+  var precipitation : Double? = nil
 
   enum CodingKeys: String, CodingKey {
 
@@ -17,7 +17,7 @@ struct Minutely: Codable {
     let values = try decoder.container(keyedBy: CodingKeys.self)
 
     dt            = try values.decodeIfPresent(Int.self , forKey: .dt            )
-    precipitation = try values.decodeIfPresent(Int.self , forKey: .precipitation )
+    precipitation = try values.decodeIfPresent(Double.self , forKey: .precipitation )
  
   }
